Добавлен пустой файл, куда будут переносится общие файл и переделано изменение статуса нагредва и двигателя по условию
